Consolidated vox_cube and blocks modules.
